solutions - Occupied bandwidth is where most (but not all) of the power of the emission lies
According to ITU recommendation ITU-R SM.328-8
1.13 Occupied bandwidth
The width of a frequency band such that, below the lower and above the upper frequency limits, the mean
powers emitted are each equal to a specified percentage β/2 of the total mean power of a given emission.
Unless otherwise specified by the Radiocommunication Assembly for the appropriate class of emission, the
value of β/2 should be taken as 0.5% (RR Article 1, No. 147).